Change to a section to mention that source cannot be Azure SQL DB for Project Files.
At the following step, we should mention specifically that the SSISDB created with Azure SQL DB, cannot be used as a source here.
On top of the page, it says all supported SQL versions, but based on our testing Azure SQL DB cannot be used as source here. We get the error as shown in screenshot below.

Section following: Line 86 of the code.
""""2. On the **Select Source** page, select the existing SSIS project to deploy.""""

Error when using SQL DB as Source. "Failed to connect to server barcelona.database.windows.net."